Refactor executionstate package
