+++ service/history/nDCWorkflow.go
@@ -37,7 +37,7 @@ import (
 	"go.temporal.io/server/common/cache"
 	"go.temporal.io/server/common/cluster"
 	"go.temporal.io/server/common/payloads"
-	"go.temporal.io/server/common/persistence/wss"
+	"go.temporal.io/server/common/persistence/executionstate"
 )
 
 type (
@@ -273,7 +273,7 @@ func (r *nDCWorkflowImpl) terminateWorkflow(
 
 func (r *nDCWorkflowImpl) zombiefyWorkflow() error {
 
-	return wss.UpdateWorkflowStateStatus(r.mutableState.GetExecutionState(),
+	return executionstate.SetStateStatus(r.mutableState.GetExecutionState(),
 		enumsspb.WORKFLOW_EXECUTION_STATE_ZOMBIE,
 		enumspb.WORKFLOW_EXECUTION_STATUS_RUNNING,
 	)
